# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "version": "1.0.0",
    "title": "Swagger for NC News Api",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "harpal.gupta@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "blah:2020",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/articles/{articleId}/comments": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get Comments for article",
        "description": "",
        "operationId": "getCommentsForArticle",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/articles/{articleId}/comments/{commentId}": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get Comment",
        "description": "",
        "operationId": "getComment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  }
}